requirements.txt
fastapi
uvicorn[standard]
pydantic
streamlit
requests
python-dateutil
